+++ API/Urb.IRepositories/ITokenManagerRepository.cs
@@ -1,0 +1,12 @@
+﻿using System;
+using System.Collections.Generic;
+using System.Linq;
+using System.Text;
+using System.Threading.Tasks;
+
+namespace Urb.Application.Urb.IRepositories
+{
+    internal interface ITokenManagerRepository
+    {
+    }
+}
