+++ Presentation/Class1.cs
@@ -1,0 +1,7 @@
+﻿namespace Presentation
+{
+    public class Class1
+    {
+
+    }
+}
