+++ Urb.Plan.v2.sln
@@ -5,9 +5,13 @@ VisualStudioVersion = 17.5.33530.505
 MinimumVisualStudioVersion = 10.0.40219.1
 Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "Urb.Plan.v2", "Urb.Plan.v2\Urb.Plan.v2.csproj", "{87007E6D-08C4-4930-94EF-D742FDC575A2}"
 EndProject
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Urb.Domain", "Urb.Domain\Urb.Domain.csproj", "{6D142A86-B281-4B0E-B311-9FE5FA83B894}"
+Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "Urb.Domain", "Urb.Domain\Urb.Domain.csproj", "{6D142A86-B281-4B0E-B311-9FE5FA83B894}"
 EndProject
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Urb.Infrastructure", "Urb.Infrastructure\Urb.Infrastructure.csproj", "{8B02DB65-8E60-4695-975A-ECAFBD8C46B0}"
+Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "Urb.Infrastructure", "Urb.Infrastructure\Urb.Infrastructure.csproj", "{8B02DB65-8E60-4695-975A-ECAFBD8C46B0}"
+EndProject
+Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "Urb.Application", "API\Urb.Application.csproj", "{7395CB59-4757-4BF0-AE5B-9B69753CAC42}"
+EndProject
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Urb.Persistance", "Urb.Persistance\Urb.Persistance.csproj", "{986DC656-588B-46DF-BA6B-68E3805B550F}"
 EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
@@ -27,6 +31,14 @@ Global
 		{8B02DB65-8E60-4695-975A-ECAFBD8C46B0}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{8B02DB65-8E60-4695-975A-ECAFBD8C46B0}.Release|Any CPU.ActiveCfg = Release|Any CPU
 		{8B02DB65-8E60-4695-975A-ECAFBD8C46B0}.Release|Any CPU.Build.0 = Release|Any CPU
+		{7395CB59-4757-4BF0-AE5B-9B69753CAC42}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{7395CB59-4757-4BF0-AE5B-9B69753CAC42}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{7395CB59-4757-4BF0-AE5B-9B69753CAC42}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{7395CB59-4757-4BF0-AE5B-9B69753CAC42}.Release|Any CPU.Build.0 = Release|Any CPU
+		{986DC656-588B-46DF-BA6B-68E3805B550F}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{986DC656-588B-46DF-BA6B-68E3805B550F}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{986DC656-588B-46DF-BA6B-68E3805B550F}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{986DC656-588B-46DF-BA6B-68E3805B550F}.Release|Any CPU.Build.0 = Release|Any CPU
 	EndGlobalSection
 	GlobalSection(SolutionProperties) = preSolution
 		HideSolutionNode = FALSE
