--- Urb.Plan.v2/Pages/_ViewImports.cshtml
@@ -1,3 +1,0 @@
-﻿@using Urb.Plan.v2
-@namespace Urb.Plan.v2.Pages
-@addTagHelper *, Microsoft.AspNetCore.Mvc.TagHelpers
