--- Urb.Plan.v2/Services/JWTService.cs
@@ -1,77 +1,0 @@
-﻿using Microsoft.AspNetCore.Identity;
-using Microsoft.EntityFrameworkCore.Metadata.Internal;
-using Microsoft.IdentityModel.Tokens;
-using Microsoft.VisualBasic;
-using System.IdentityModel.Tokens.Jwt;
-using System.Security.Claims;
-using System.Text;
-using Urb.Domain.Urb.Models;
-using Urb.Plan.v2.IServices;
-
-namespace Urb.Plan.v2.Services
-{
-    public class JWTService: IJWTService
-    {
-
-        private readonly AppSettings _appSettings;
-
-        public JWTService (AppSettings appSettings)
-        {
-            _appSettings = appSettings;
-        }
-
-        public string GenerateToken(IdentityUser user)
-        {
-            var tokenHandler = new JwtSecurityTokenHandler();
-            var tokenKey = Encoding.ASCII.GetBytes(_appSettings.JWTKey);
-
-            var tokenDescriptor = new SecurityTokenDescriptor
-            {
-                Subject = new ClaimsIdentity(new[] { new Claim("id", user.Id.ToString()) }),
-                Expires = DateTime.UtcNow.AddDays(7),
-                SigningCredentials = new SigningCredentials(new SymmetricSecurityKey(tokenKey), SecurityAlgorithms.HmacSha256Signature)
-            };
-
-            var token = tokenHandler.CreateToken(tokenDescriptor);
-            return tokenHandler.WriteToken(token);
-        }
-        public int? ValidateToken (string token)
-        {
-            if(token == null)
-            {
-                return null;
-            }
-
-            AppSettings appSettings = new AppSettings();
-            JwtSecurityTokenHandler jwtSecurityTokenHandler = new JwtSecurityTokenHandler();
-            var tokenKey = Encoding.ASCII.GetBytes(appSettings.JWTKey);
-
-            try
-            {
-                jwtSecurityTokenHandler.ValidateToken(token, new TokenValidationParameters
-                {
-                    ValidateIssuerSigningKey = true,
-                    IssuerSigningKey = new SymmetricSecurityKey(tokenKey),
-                    ValidateIssuer = false,
-                    ValidateAudience = false,
-                    ClockSkew = TimeSpan.Zero
-
-                }, out SecurityToken validatedToken);
-
-                var jwtToken = (JwtSecurityToken)validatedToken;
-
-                var userId = int.Parse(jwtToken.Claims.First(x => x.Type == "id").Value);               
-
-                return userId;
-            }
-
-            catch
-            {
-                return null;
-            }
-
-
-                
-        }
-    }
-}
